chore: remove unnecessary part of build.gradle file
Context: This build.gradle file wasn't even being used before because it wasn't referenced in the plugin.xml. Now, we just want to use it to declare the dependencies for the Firebase Android SDK for Dynamic Links, so we can remove this part.

References: https://outsystemsrd.atlassian.net/browse/RMET-3608